adds supplier address deletion

+++ src/Adapter/Supplier/CommandHandler/DeleteSupplierHandler.php
@@ -26,10 +26,14 @@
 
 namespace PrestaShop\PrestaShop\Adapter\Supplier\CommandHandler;
 
+use Address;
 use PrestaShop\PrestaShop\Adapter\Entity\Db;
+use PrestaShop\PrestaShop\Adapter\Supplier\SupplierAddressProvider;
 use PrestaShop\PrestaShop\Adapter\Supplier\SupplierOrderValidator;
+use PrestaShop\PrestaShop\Adapter\Validate;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Command\DeleteSupplierCommand;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\CommandHandler\DeleteSupplierHandlerInterface;
+use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\CannotDeleteSupplierAddressException;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\CannotDeleteSupplierException;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\CannotDeleteSupplierProductRelationException;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\SupplierException;
@@ -54,15 +58,23 @@ final class DeleteSupplierHandler implements DeleteSupplierHandlerInterface
     private $dbPrefix;
 
     /**
+     * @var SupplierAddressProvider
+     */
+    private $supplierAddressProvider;
+
+    /**
      * @param SupplierOrderValidator $supplierOrderValidator
+     * @param SupplierAddressProvider $supplierAddressProvider
      * @param string $dbPrefix
      */
     public function __construct(
         SupplierOrderValidator $supplierOrderValidator,
+        SupplierAddressProvider $supplierAddressProvider,
         $dbPrefix
     ) {
         $this->supplierOrderValidator = $supplierOrderValidator;
         $this->dbPrefix = $dbPrefix;
+        $this->supplierAddressProvider = $supplierAddressProvider;
     }
 
     /**
@@ -107,6 +119,16 @@ final class DeleteSupplierHandler implements DeleteSupplierHandlerInterface
                 );
             }
 
+            if (false === $this->deleteSupplierAddress($command->getSupplierId())) {
+                Db::getInstance()->execute('ROLLBACK;');
+                throw new CannotDeleteSupplierAddressException(
+                    sprintf(
+                        'Unable to set deleted flag for supplier with id "%s" address',
+                        $command->getSupplierId()->getValue()
+                    )
+                );
+            }
+
             if (false === $entity->delete()) {
                 Db::getInstance()->execute('ROLLBACK;');
                 throw new CannotDeleteSupplierException(
@@ -132,7 +154,7 @@ final class DeleteSupplierHandler implements DeleteSupplierHandlerInterface
     }
 
     /**
-     * Deletes product supplier relation
+     * Deletes product supplier relation.
      *
      * @param SupplierId $supplierId
      *
@@ -143,5 +165,26 @@ final class DeleteSupplierHandler implements DeleteSupplierHandlerInterface
         $sql = 'DELETE FROM `' . $this->dbPrefix . 'product_supplier` WHERE `id_supplier`=' . $supplierId->getValue();
 
         return Db::getInstance()->execute($sql);
+    }
+
+    /**
+     * Deletes supplier address.
+     *
+     * @param SupplierId $supplierId
+     *
+     * @return bool
+     */
+    private function deleteSupplierAddress(SupplierId $supplierId)
+    {
+        $supplierAddressId = $this->supplierAddressProvider->getIdBySupplier($supplierId->getValue());
+
+        $address = new Address($supplierAddressId);
+
+        if ($address->id) {
+            $address->deleted = true;
+            return $address->update();
+        }
+
+        return true;
     }
 }
